LayoutTests/fast/forms/input-maxlength-paste-clusters-in-middle-expected.txt

This test attempts to paste text in the middle of a text field with maxlength.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.value is "abc"
PASS input.getAttribute("maxlength") is "3"
input.selectionStart = input.selectionEnd = 2
getSelection().modify("extend", "backward", "character")
PASS document.execCommand("insertText", false, "स्"); input.value is "aस्c"
PASS input.value.length is 4
PASS successfullyParsed is true

TEST COMPLETE
